requirement.txt
torch
torchvision
numpy
roma
gradio
matplotlib
tqdm
opencv-python
scipy
evo
einops
trimesh
tensorboard
pyglet
pillow
gdown
imageio
pyquaternion
wandb